inventory:refactor: refactor entity

+++ inventory_ms/domain/src/entities/product_entity.rs
@@ -1,3 +1,4 @@
+#[derive(Default)]
 pub struct ProductEntity {
     pub id: String,
     pub product_category: String,
@@ -5,9 +6,9 @@ pub struct ProductEntity {
     pub title: String,
     pub subtitle: String,
     pub authors: Vec<String>,
-    pub amount_in_stock: i32,
+    pub amount_in_stock: i64,
     pub created_at: String,
     pub updated_at: String,
-    pub num_pages: i32,
+    pub num_pages: i64,
     pub tags: Vec<String>,
 }
